Refactor image retrieval (#320)
This makes the image retrieval process not rely on a FutureBuilder.
Previously we recreated Image behavior by showing a loader whilst the image loads, using a completer to detect when the image had finished load.
Now we rely only on the good old state and some ifs in the build method.
Also, we have split that code from the custom child code. The main widget had logics for both image and custom child modes. Now we split that into two internal widgets, the wrappers.

This should resolve the following issues #316 #303