Added pageSnapping option (#542)
Adds the pageSnapping option to give the dev the option to view
continually the image from the gallery, without snapping every time.

The main reason for me doing this addition is to add this option also on
the [pdfx](https://pub.dev/packages/pdfx) package, it is useful for the
web apps with vertical viewing.